Improve code reuse for house checks
Pattern checks for a house are a little trickier than for a row or
column because the cells in a house form a 3x3 grid instead of a 1x9
vector.  There is also inconsistency between rows and columns because we
need to know to flip the order of the coordinates.

We can get better code reuse for group checks by storing a 1x9 vector of
cells for each group during solver initialization.  We also store the
cell's row and column in the `Cell` struct so that we can easily
retrieve its board location.