Short-circuit checks if candidates are invalidated
If a check starts by selecting a set of viable candidates that meet a
certain set of requirements, then any change to the board state could
invalidate the remaining candidates.  This could result in incorrect
matches, so it is best to short-circuit after the first successful
candidate.  Otherwise, we would need to keep re-checking the criteria
for each candidate, which would complicate the code and counteract any
performance advantage we might have gained by checking all the
candidates in one loop.

After short-circuiting, we can either continue with any additional
checks which select a fresh set of candidates, or we can just return
to the start of the solver loop.  Any candidates that were skipped
will be picked up in a later solver pass if still needed.