Just use the basic `int` type for digit values
Using the `int8` type adds complication to the code because of the
required conversions to and from `int` for many standard library
functions, and requires separate instantations of generic types and
functions.  The little bit of memory saved by using the shorter type
is unlikely to out-weigh the additional complexity, especially after
taking memory alignment rules into account.