Print solved puzzle in a compact form
When printing the puzzle, we don't always need to show the full pencil
mark grid, e.g. when the puzzle has been fully solved or when showing
the initial puzzle.  In these cases, we can print the board in a much
more compact form that doesn't take up so many screen lines.