Handle cell candidates in `puzzle` package
If we handle removing conflicting candidates in the `puzzle` package as
soon as a solved (or given) value is placed, then it allows us to do
things like checking for an invalid puzzle while reading or printing
out a pencil mark grid for the initial puzzle state.  This only affects
the cell candidates that are stored in the Puzzle struct, not cached
candidates that are only needed by the Solver.

However, since the cell candidates are removed as soon as the value is
placed into the grid, the Solver can no longer use the presence of a
candidate as a check to ensure that each cell is only processed once.
Instead, the cached locations must be used to find peer cells where the
value is still a valid candidate.  This makes the code a bit more
complicated, but in theory it should also be slightly faster because we
are only processing cells that are known to contain the candidate.