Isolate terminal output from solving process
Added an Options structure for the solver to enable or disable live
progress updates that are mostly used for troubleshooting the solver.
Solution steps are now collected into a list that can be printed by the
driver application after the solver is finished.  This facilitates
proper benchmarking of the solver without measuring extra overhead from
the terminal output.