Implement bitset for better performance
All of the sets we use have only 9 possible values; [0,8] for location
indexes or [1,9] for digit values.  For these small integer values, we
can store the whole set as a bit mask in a 16-bit unsigned integer, and
all set operations can be implemented with very efficient "bit
twiddling" operations.  This one simple change allows the solver to run
in about 1/3 of the time on average.