Change how checks are counted
Incrementing the check count only once per technique checked doesn't
give any indication of how many candidates need to be checked to
determine a technique was matched, which is a major contributor to
overall performance.  This change increments the count for each
"decision" that is made, which makes it possible to compare whether a
change to the implementation of a technique causes the number of checks
to increase or decrease.

Note that every decision does not have the same cost in terms of
performance, so the check counts should not be viewed as an absolute
measurement.  The purpose of the check count is to help determine if
there has been an increase or decrease in the amount of work that needs
to be done to recognize particular techniques.